services/transaction-processor/requirements.txt
bcrypt==4.3.0
boto==2.49.0
cffi==2.0.0
cryptography==46.0.1
invoke==2.2.0
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.0
python-dotenv==1.0.0
mysql-connector-python==9.4.0
